requirements.txt
PyQt6>=6.6.0
psutil>=5.9.0
numpy>=1.26.0
sounddevice>=0.5.0
pyttsx3>=2.90
groq>=0.4.0
mss>=10.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
